requirements.txt
asgiref==3.7.2
cx_Oracle==8.3.0
Django==5.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
whitenoise==6.6.0
